Fix race condition in capturing checkpoints
Summary:
When multiple clients start up at once when shards are already claimed,
two clients may claim the same shard as they enter the pool. They
will both start trying to capture the shard, and once the shard is released
by the owner, there is a race where they can both capture it, resulting
in duplicated data until one of them releases it in the next refreshShards.

Using a conditional expression seems to introduce a latency of 2-3 seconds
between checkpoint release and capturing. That's ok for now with a minute
ticker on refreshShards, but we may want to improve it if we decrease our SLAs.

Also update unit test to use multiple clients.

Test Plan: Ran unit tests and spade integration test.

Reviewers: jackgao, spencer

Reviewed By: spencer

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D3817